+++ src/main/java/com/bettercontent/bettercontentfixes/performance/DistantGenerationControl.java
@@ -1,0 +1,15 @@
+package com.bettercontent.bettercontentfixes.performance;
+
+public interface DistantGenerationControl {
+    boolean isAvailable();
+
+    boolean baseEnabled();
+
+    boolean overrideActive();
+
+    boolean pause();
+
+    void resume();
+
+    String description();
+}